+++ audionimbus/src/effect/ambisonics/decode.rs
@@ -73,6 +73,24 @@ impl AmbisonicsDecodeEffect {
         .into()
     }
 
+    /// Retrieves a single frame of tail samples from an Ambisonics decode effect’s internal buffers.
+    ///
+    /// After the input to the Ambisonics decode effect has stopped, this function must be called instead of [`Self::apply`] until the return value indicates that no more tail samples remain.
+    ///
+    /// The output audio buffer must have as many channels as needed for the speaker layout specified when creating the effect (if using panning) or 2 channels (if using binaural rendering).
+    pub fn tail<O>(&self, output_buffer: &AudioBuffer<O>) -> AudioEffectState
+    where
+        O: AsRef<[Sample]> + AsMut<[Sample]>,
+    {
+        unsafe {
+            audionimbus_sys::iplAmbisonicsDecodeEffectGetTail(
+                self.raw_ptr(),
+                &mut *output_buffer.as_ffi(),
+            )
+        }
+        .into()
+    }
+
     /// Returns the number of tail samples remaining in an Ambisonics decode effect’s internal buffers.
     ///
     /// Tail samples are audio samples that should be played even after the input to the effect has stopped playing and no further input samples are available.