+++ audionimbus/src/effect/ambisonics/rotation.rs
@@ -79,6 +79,24 @@ impl AmbisonicsRotationEffect {
         .into()
     }
 
+    /// Retrieves a single frame of tail samples from an Ambisonics rotation effect’s internal buffers.
+    ///
+    /// After the input to the Ambisonics rotation effect has stopped, this function must be called instead of [`Self::apply`] until the return value indicates that no more tail samples remain.
+    ///
+    /// The output audio buffer must have as many channels as needed for the Ambisonics order specified when creating the effect.
+    pub fn tail<O>(&self, output_buffer: &AudioBuffer<O>) -> AudioEffectState
+    where
+        O: AsRef<[Sample]> + AsMut<[Sample]>,
+    {
+        unsafe {
+            audionimbus_sys::iplAmbisonicsRotationEffectGetTail(
+                self.raw_ptr(),
+                &mut *output_buffer.as_ffi(),
+            )
+        }
+        .into()
+    }
+
     /// Returns the number of tail samples remaining in an Ambisonics rotation effect’s internal buffers.
     ///
     /// Tail samples are audio samples that should be played even after the input to the effect has stopped playing and no further input samples are available.